pyproject.toml
[tool.poetry]
name = "spotify-now-playing-telegram"
version = "0.1.0"
description = ""
authors = ["Marco Burro <marcoburro98@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
tornado = "6.4.1"
pony = "0.7.14"
pyfy = "2.0.8"
python-telegram-bot = "13.5"
psycopg2-binary = "2.8.6"
python-dotenv = "0.17.1"
urllib3 = "1.25.9"
python-ffmpeg = "^2.0.12"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
